moved utility class to corresponding product package
